Change to ensure defaults are set when partial values are supplied
With just one call options like POINT are not initialized properly
unless values are actually set in the main program for the values.

So if you did not declare a value for POINT in the examples like

   REAL :: POINT(3)=[10,20,30]

but just set the values in the prototype

   CMD='-point 10,20,30'

the called only changing the second value

   ./demo1 -point ,100

instead of POINT=[10,100,30] it would have undefined values like
POINT=[NaN,100,NaN].

And since NAMELIST cannot be passed decided best approach was to
call twice, once to use prototype to set defaults, again to use
command line values.