Added option to commandline(3f) to take a NAMELIST group name
Also added demo6 as an example of how the NAME= argument can be
used to allow different command options. This can be used for
wanting different command behaviors depending on the name the
executable is called with, for creating a command with subcommands,
or perhaps changing behavior based on an external condition like
and environment variable.

PROGRAMS/demo6.f90 shows a rather long example of how to create
a command with subcommands, so you can enter

    demo6 run [options for running]
    demo6 test [options for testing]
    demo6 --help
    demo6 run --help

like commands like git(1) do.

and so on.